Moved RegisterApp to component & Dialog box Added cancel button & cancel props to registerApp
