+++ src/views/dashboard.tsx
@@ -1,11 +1,11 @@
-import { Text, Stack, Persona, PersonaSize, Icon, Link } from "office-ui-fabric-react";
+import { Text, Stack, Persona, PersonaSize, Icon, Link, Dialog, DialogType, DefaultButton } from "office-ui-fabric-react";
 import React from "react";
 import { GetUserAvatar, GetCurrentUser, IDiscordUser, discordAuthEndpoint } from "../common/services/discord";
 
 import HoverBox from "../components/HoverBox";
 import styled from "styled-components";
 import { NavLink } from "react-router-dom";
-import { RegisterApp } from "./dashboard/registerapp";
+import { RegisterApp } from "../components/registerApp";
 
 const DashboardHeader = styled.header`
 background: linear-gradient(to bottom,#005799 0,#0076d1);
@@ -18,6 +18,8 @@ padding: 15px 0px;
 export const Dashboard = () => {
     const [welcomeMessage, setWelcomeMessage] = React.useState("Signing in...");
     const [userIcon, setUserIcon] = React.useState("");
+
+    const [appRegistrationShown, setAppRegistrationShown] = React.useState(false);
 
     React.useEffect(() => {
         setupLoggedInUser();
@@ -58,12 +60,13 @@ export const Dashboard = () => {
                     <PersonaDark size={PersonaSize.extraLarge} text={welcomeMessage} imageUrl={userIcon} />
 
                     <Stack horizontal wrap verticalAlign="end" tokens={{ childrenGap: 10 }} style={{ marginLeft: 10 }}>
-                        <NavLink style={{ color: "white", width: "150px", textDecoration: "none" }} to="/dashboard/registerapp">
+
+                        <Link style={{ color: "white", width: "150px", textDecoration: "none" }} onClick={() => setAppRegistrationShown(true)}>
                             <Stack verticalAlign="end" horizontalAlign="center" tokens={{ childrenGap: 5 }}>
                                 <Icon style={{ fontSize: 35 }} iconName="AppIconDefaultAdd"></Icon>
                                 <Text variant="mediumPlus">Register an app</Text>
                             </Stack>
-                        </NavLink>
+                        </Link>
 
                         <Link style={{ color: "white", width: "150px", textDecoration: "none", marginBottom: 2.5 }} to="/dashboard/registerapp">
                             <Stack verticalAlign="end" horizontalAlign="center" tokens={{ childrenGap: 5 }}>
@@ -71,6 +74,7 @@ export const Dashboard = () => {
                                 <Text variant="mediumPlus">Manage your roles</Text>
                             </Stack>
                         </Link>
+
                     </Stack>
 
                     <DashboardColumnFiller style={{ width: 200 }} />
@@ -100,14 +104,12 @@ export const Dashboard = () => {
                     </HoverBox>
                 </Stack>
 
-
-                <Stack horizontalAlign="center" tokens={{ childrenGap: 10 }}>
-                    <Stack horizontal tokens={{ childrenGap: 10 }}>
-                        <Icon style={{ fontSize: 28 }} iconName="AppIconDefaultAdd" />
-                        <Text variant="xLarge">Register an app</Text>
-                    </Stack>
-                    <RegisterApp />
-                </Stack>
+                <Dialog isOpen={appRegistrationShown} dialogContentProps={{
+                    type: DialogType.largeHeader,
+                    title: 'Register an app',
+                }}>
+                    <RegisterApp onCancel={() => setAppRegistrationShown(false)} />
+                </Dialog>
             </Stack>
         </Stack >
     )
